// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      orders: {
        Row: {
          amount: number
          country: string | null
          created_at: string
          currency: string
          email: string
          id: string
          ip: string | null
          is_new_customer: boolean
          order_ref: string
          platform: string
          reasons: string[]
          risk_score: number
          status: Database["public"]["Enums"]["risk_status"]
          user_id: string
        }
        Insert: {
          amount: number
          country?: string | null
          created_at?: string
          currency?: string
          email: string
          id?: string
          ip?: string | null
          is_new_customer?: boolean
          order_ref: string
          platform: string
          reasons?: string[]
          risk_score?: number
          status?: Database["public"]["Enums"]["risk_status"]
          user_id: string
        }
        Update: {
          amount?: number
          country?: string | null
          created_at?: string
          currency?: string
          email?: string
          id?: string
          ip?: string | null
          is_new_customer?: boolean
          order_ref?: string
          platform?: string
          reasons?: string[]
          risk_score?: number
          status?: Database["public"]["Enums"]["risk_status"]
          user_id?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never